Add files via upload
The objective of this project is to predict the future share values by using the previous data of the company. This project is developed using the Long-Short-Term-Memory model and the AAPL.csv file which includes all the information of the company. In this project, I have predicted the next 30days share values.